+++ level.go
@@ -21,47 +21,6 @@ const (
 	DebugLevel
 )
 
-var (
-	// without colors
-	erroText = "[ERRO]"
-	warnText = "[WARN]"
-	infoText = "[INFO]"
-	dbugText = "[DBUG]"
-	// with colors
-	erro = pio.Red(erroText)
-	warn = pio.Purple(warnText)
-	info = pio.LightGreen(infoText)
-	dbug = pio.Yellow(dbugText)
-)
-
-// returns a [PREFIX] based on the "level" and "enableColor".
-func prefixFromLevel(level Level, enableColor bool) string {
-	switch level {
-	case ErrorLevel:
-		if !enableColor {
-			return erroText
-		}
-		return erro
-	case WarnLevel:
-		if !enableColor {
-			return warnText
-		}
-		return warn
-	case InfoLevel:
-		if !enableColor {
-			return infoText
-		}
-		return info
-	case DebugLevel:
-		if !enableColor {
-			return dbugText
-		}
-		return dbug
-	default:
-		return ""
-	}
-}
-
 func fromLevelName(levelName string) Level {
 	switch levelName {
 	case "error":
@@ -78,3 +37,153 @@ func fromLevelName(levelName string) Level {
 		return DisableLevel
 	}
 }
+
+var (
+
+	// Author's note:
+	// Here I choose to apply the below pattern of modifying the raw text and the colorful text
+	// for performance reasons, we could just create the GetTextForLevel to generate
+	// the colors on-the-fly or add a sync.Once but both of these would reduce our performance.
+
+	errorText = "[ERRO]"
+	// errorTextWithColor is the color that will be printed
+	// when Error/Errorf functions are being used, if `Printer#IsTerminal` is true.
+	//
+	// Defaults to a red color.
+	errorTextWithColor = pio.Red(errorText)
+	// ErrorText can modify the prefix that will be prepended
+	// to the output message log when `Error/Errorf` functions are being used.
+	//
+	// If "newRawText" is empty then it will just return the current prefix string value.
+	// If "newColorfulText" is empty then it will update the text color version using
+	// the default values by using the new raw text.
+	//
+	// Defaults to "[ERRO]" and pio.Red("[ERRO]").
+	ErrorText = func(newRawText string, newColorfulText string) (oldRawText string) {
+		oldRawText = errorText
+		if newRawText != "" {
+			errorText = newRawText
+			errorTextWithColor = pio.Red(newRawText)
+		}
+		if newColorfulText != "" {
+			errorTextWithColor = newColorfulText
+		}
+		return
+	}
+
+	warnText = "[WARN]"
+	// warnTextWithColor is the color that will be printed
+	// when Warn/Warnf functions are being used, if `Printer#IsTerminal` is true.
+	//
+	// Defaults to a purplish color.
+	warnTextWithColor = pio.Purple(warnText)
+	// WarnText can modify the prefix that will be prepended
+	// to the output message log when `Warn/Warnf` functions are being used.
+	//
+	// If "newRawText" is empty then it will just return the current prefix string value.
+	// If "newColorfulText" is empty then it will update the text color version using
+	// the default values by using the new raw text.
+	//
+	// Defaults to "[WARN]" and pio.Purple("[WARN]").
+	WarnText = func(newRawText string, newColorfulText string) (oldRawText string) {
+		oldRawText = warnText
+		if newRawText != "" {
+			warnText = newRawText
+			warnTextWithColor = pio.Purple(newRawText)
+		}
+		if newColorfulText != "" {
+			warnTextWithColor = newColorfulText
+		}
+		return
+	}
+
+	infoText = "[INFO]"
+	// infoTextWithColor is the color that will be printed
+	// when Info/Infof functions are being used, if `Printer#IsTerminal` is true.
+	//
+	// Defaults to a mix of light green and blue color.
+	infoTextWithColor = pio.LightGreen(infoText)
+	// InfoText can modify the prefix that will be prepended
+	// to the output message log when `Info/Infof` functions are being used.
+	//
+	// If "newRawText" is empty then it will just return the current prefix string value.
+	// If "newColorfulText" is empty then it will update the text color version using
+	// the default values by using the new raw text.
+	//
+	// Defaults to "[INFO]" and pio.LightGreen("[INFO]").
+	InfoText = func(newRawText string, newColorfulText string) (oldRawText string) {
+		oldRawText = infoText
+		if newRawText != "" {
+			infoText = newRawText
+			infoTextWithColor = pio.LightGreen(newRawText)
+		}
+		if newColorfulText != "" {
+			infoTextWithColor = newColorfulText
+		}
+		return
+	}
+
+	debugText = "[DBUG]"
+	// debugTextWithColor is the color that will be printed
+	// when Debug/Debugf functions are beingused, if `Printer#IsTerminal` is true.
+	//
+	// Defaults to a yellow color.
+	debugTextWithColor = pio.Yellow(debugText)
+	// DebugText can modify the prefix that will be prepended
+	// to the output message log when `Info/Infof` functions are being used.
+	//
+	// If "newRawText" is empty then it will just return the current prefix string value.
+	// If "newColorfulText" is empty then it will update the text color version using
+	// the default values by using the new raw text.
+	//
+	// Defaults to "[DBUG]" and pio.Yellow("[DBUG]").
+	DebugText = func(newRawText string, newColorfulText string) (oldRawText string) {
+		oldRawText = debugText
+		if newRawText != "" {
+			debugText = newRawText
+			debugTextWithColor = pio.Yellow(newRawText)
+		}
+		if newColorfulText != "" {
+			debugTextWithColor = newColorfulText
+		}
+		return
+	}
+
+	// GetTextForLevel is the function which
+	// has the "final" responsibility to generate the text (colorful or not)
+	// that is prepended to the leveled log message
+	// when `Error/Errorf, Warn/Warnf, Info/Infof or Debug/Debugf`
+	// functions are being called.
+	//
+	// It can be used to override the default behavior, at the start-up state.
+	GetTextForLevel = func(level Level, enableColor bool) string {
+		switch level {
+		case ErrorLevel:
+			if !enableColor {
+				return errorText
+			}
+			return errorTextWithColor
+
+		case WarnLevel:
+			if !enableColor {
+				return warnText
+			}
+			return warnTextWithColor
+
+		case InfoLevel:
+			if !enableColor {
+				return infoText
+			}
+			return infoTextWithColor
+
+		case DebugLevel:
+			if !enableColor {
+				return debugText
+			}
+			return debugTextWithColor
+
+		default:
+			return ""
+		}
+	}
+)